pyproject.toml
[project]
name = "sozo"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "~=3.12"
dependencies = ["flask>=3.0.3", "protobuf>=5.28.3", "pyzmq>=26.2.0"]

[tool.uv]
package = false

[tool.ruff]
line-length = 100

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []
ignore = ["B905", "D417", "E501", "PLR0913"]

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"
